container/decision_trees/requirements.txt
numpy==1.19.2
scipy==1.5.3
scikit-learn==0.23.2
pandas==1.1.3
flask==1.1.2
gevent==20.9.0
gunicorn==20.0.4
typer==0.3.2
fastapi==0.61.1
pyyaml==5.3.1
gunicorn==20.0.4
uvicorn==0.12.2